// source/server/types/inspection.ts
/**
 * 3D Foundation Project
 * Copyright 2023 Smithsonian Institution
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface IMeshSmithInspection
{
    filePath: string;
    meshes: {
        geometry: {
            boundingBox: {
                min: number[];
                max: number[];
            };
            center: number[];
            size: number[];
        };
        statistics: {
            hasBones: boolean;
            hasNormals: boolean;
            hasTangentsAndBitangents: boolean;
            hasTexCoords: boolean;
            hasVertexColors: boolean;
            numVertices: number;
            numFaces: number;
            numColorChannels: number;
            numTexCoordChannels: number;
        };
    }[];
    scene: {
        geometry: {
            boundingBox: {
                min: number[];
                max: number[];
            };
            center: number[];
            size: number[];
        };
        statistics: {
            numAnimations: number;
            numCameras: number;
            numLights: number;
            numMaterials: number;
            numTextures: number;
            numMeshes: number;
            numVertices: number;
            numFaces: number;
        };
    }
    type: "report";
}